requirements.txt
pillow==12.1.1
pyaes==1.6.1
pyasn1==0.6.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dotenv==1.2.1
qasync==0.28.0
qrcode==8.2
rsa==4.9.1
Telethon==1.42.0
